% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_api.R
\name{as.data.frame.finbif_api}
\alias{as.data.frame.finbif_api}
\title{Coerce \code{finbif_api} object to a \code{data.frame}}
\usage{
\method{as.data.frame}{finbif_api}(x, ...)
}
\arguments{
\item{x}{A \code{finbif_api} object.}

\item{...}{Additional arguments. Not used.}
}
\value{
\code{data.frame}
}
\description{
Converts the result of a FinBIF query to a \code{data.frame}.
}
