% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.data.frame.finbif_records}
\alias{as.data.frame.finbif_records}
\alias{as.data.frame.finbif_records_list}
\title{Coerce a \code{finbif_records*} object to a \code{data.frame}}
\usage{
\method{as.data.frame}{finbif_records}(x, ...)

\method{as.data.frame}{finbif_records_list}(x, ..., quiet = FALSE)
}
\arguments{
\item{x}{A \code{finbif_records*} object.}

\item{...}{Additional arguments. Not used.}

\item{quiet}{Logical. If \code{TRUE} (default) suppress progress indicator of
conversion.}
}
\value{
A \code{data.frame}.
}
\description{
Converts the result of a FinBIF query to a \code{data.frame}.
}
\examples{
\dontrun{

# Download the latest records from FinBIF
# and convert to a `data.frame`
resp <- finbif_records()
df <- as.data.frame(resp)
}
}
