% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_admin_status.R
\name{finbif_admin_status}
\alias{finbif_admin_status}
\title{FinBIF administrative status}
\usage{
finbif_admin_status()
}
\value{
A data.frame.
}
\description{
Display the administrative statuses used in the FinBIF database.
}
\examples{
\dontrun{

# Display the administrative statuses used by FinBIF
finbif_admin_status()
}
}
