% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_check_taxa.R
\name{finbif_check_taxa}
\alias{finbif_check_taxa}
\title{Check FinBIF taxa}
\usage{
finbif_check_taxa(taxa)
}
\arguments{
\item{taxa}{Named list of vectors. Each vector should have the name of a
taxonomic rank (genus, species, etc.,). The elements of the vectors should
be the taxa to check and be of type character.}
}
\value{
A list.
}
\description{
Check that taxa are in the FinBIF database.
}
