% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_collections.R
\name{finbif_collections}
\alias{finbif_collections}
\title{FinBIF collections}
\usage{
finbif_collections(subcollections = FALSE)
}
\arguments{
\item{subcollections}{Logical. Display subcollections as well the main
collections.}
}
\value{
A data.frame.
}
\description{
Display information on the collections in the FinBIF database.
}
\examples{
\dontrun{

# Display the informal taxonomic groups used by FinBIF
finbif_collections()
}
}
