% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_countries.R
\name{finbif_countries}
\alias{finbif_countries}
\title{Countries}
\usage{
finbif_countries()
}
\value{
A data.frame.
}
\description{
Display the country names and codes used in the FinBIF database.
}
\examples{
\dontrun{

# Display the country names and codes used in FinBIF
finbif_countries()
}
}
