% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_finnish_occurrence.R
\name{finbif_finnish_occurrence}
\alias{finbif_finnish_occurrence}
\title{Taxon Finnish occurrence statuses}
\usage{
finbif_finnish_occurrence()
}
\value{
A data.frame.
}
\description{
Display the description and codes for Finnish occurrence status used in the
FinBIF database.
}
\examples{
\dontrun{

# Display countries, provinces or municipalities
finbif_finnish_occurrence()
}
}
