% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_habitat_types.R
\name{finbif_habitat_types}
\alias{finbif_habitat_types}
\alias{finbif_habitat_qualifiers}
\title{FinBIF habitat types}
\usage{
finbif_habitat_types()

finbif_habitat_qualifiers()
}
\value{
A data.frame.
}
\description{
Display the habitat types and specific habitat types used in the FinBIF
database.
}
\examples{
\dontrun{

# Display the administrative statuses used by FinBIF
finbif_habitat_types()
finbif_habitat_qualifiers()
}
}
