% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_informal_groups.R
\name{finbif_informal_groups}
\alias{finbif_informal_groups}
\alias{fb_informal_groups}
\title{FinBIF informal groups}
\usage{
finbif_informal_groups(
  group,
  limit = 50,
  quiet = FALSE,
  locale = getOption("finbif_locale")
)
}
\arguments{
\item{group}{Character. Optional, if supplied only display this group and
its subgroups.}

\item{limit}{Integer. The maximum number informal groups to display.}

\item{quiet}{Logical. Return informal group names without displaying them.}

\item{locale}{Character. One of the supported two-letter ISO 639-1 language
codes. Current supported languages are English, Finnish and Swedish. For
data where more than one language is available the language denoted by
\code{locale} will be preferred while falling back to the other languages in the
order indicated above.}
}
\value{
A character vector (invisibly).
}
\description{
Display the informal taxonomic groups used in the FinBIF database.
}
\examples{
\dontrun{

# Display the informal taxonomic groups used by FinBIF
finbif_informal_groups()

}
}
