% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_occurrence.R
\name{finbif_last_mod}
\alias{finbif_last_mod}
\alias{fb_last_mod}
\title{Get last modified date for FinBIF occurrence records}
\usage{
finbif_last_mod(..., filter)
}
\arguments{
\item{...}{Character vectors or list of character vectors. Taxa of records
to download.}

\item{filter}{List of named character vectors. Filters to apply to records.}
}
\value{
A \code{Date} object
}
\description{
Get last modified date for filtered occurrence data from FinBIF.
}
\examples{
\dontrun{

# Get last modified date for Whooper Swan occurrence records from Finland
finbif_last_mod("Cygnus cygnus", filter = c(country = "Finland"))

}
}
