% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_life_stage.R
\name{finbif_life_stages}
\alias{finbif_life_stages}
\title{FinBIF life stages}
\usage{
finbif_life_stages()
}
\value{
A character vector.
}
\description{
Display organism life stages used in the FinBIF database.
}
\examples{
\dontrun{

# Display list of life stages
finbif_life_stages()
}
}
