% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_record_basis.R
\name{finbif_record_basis}
\alias{finbif_record_basis}
\title{FinBIF record basis types}
\usage{
finbif_record_basis()
}
\value{
A data.frame.
}
\description{
Display information on the basis of FinBIF records.
}
\examples{
\dontrun{

# Display record basis metadata
finbif_record_basis()
}
}
