% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_records.R
\name{finbif_records}
\alias{finbif_records}
\title{Get FinBIF records}
\usage{
finbif_records(fields, filters = NULL)
}
\arguments{
\item{fields}{Character vector. Columns to return.}

\item{filters}{List of named character vectors. Filters to apply to records.}
}
\value{
Data.frame.
}
\description{
Download filtered records from FinBIF.
}
