% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_red_list.R
\name{finbif_red_list}
\alias{finbif_red_list}
\title{IUCN red list}
\usage{
finbif_red_list()
}
\value{
A data.frame.
}
\description{
Display the IUCN red list statuses used in the FinBIF database.
}
\examples{
\dontrun{

# Display the IUCN red list statuses used by FinBIF
finbif_red_list()
}
}
