% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_restriction_levels.R
\name{finbif_restriction_levels}
\alias{finbif_restriction_levels}
\title{Data restriction levels}
\usage{
finbif_restriction_levels()
}
\value{
A data.frame.
}
\description{
Display the levels of data security restrictions in the FinBIF database.
}
\examples{
\dontrun{

# Display data restriction levels
finbif_restriction_levels()
}
}
