% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_restriction_reasons.R
\name{finbif_restriction_reasons}
\alias{finbif_restriction_reasons}
\title{Data restriction reasons}
\usage{
finbif_restriction_reasons()
}
\value{
A data.frame.
}
\description{
Display the reasons data in the FinBIF database have security restrictions.
}
\examples{
\dontrun{

# Display data restriction reasons
finbif_restriction_reasons()
}
}
