% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_sex_categories.R
\name{finbif_sex_categories}
\alias{finbif_sex_categories}
\title{FinBIF sex categories}
\usage{
finbif_sex_categories()
}
\value{
A data.frame.
}
\description{
Display organism sex categories used in the FinBIF database.
}
\examples{
\dontrun{

# Display list of sex categories and codes
finbif_sex_categories()
}
}
