% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_sources.R
\name{finbif_sources}
\alias{finbif_sources}
\title{FinBIF data sources}
\usage{
finbif_sources()
}
\value{
A data.frame.
}
\description{
Display the information system data sources used by FinBIF.
}
\examples{
\dontrun{

# Display data sources
finbif_sources()
}
}
