% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_taxon_ranks.R
\name{finbif_taxon_ranks}
\alias{finbif_taxon_ranks}
\title{FinBIF taxonomic ranks}
\usage{
finbif_taxon_ranks()
}
\value{
A data.frame.
}
\description{
Display taxonomic ranks used in the FinBIF database.
}
\examples{
\dontrun{

# Display the taxonomic ranks used by FinBIF
finbif_taxon_ranks()
}
}
