% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_dwc}
\alias{to_dwc}
\alias{to_native}
\title{Convert variable names to and from Darwin Core style}
\usage{
to_dwc(...)

to_native(...)
}
\arguments{
\item{...}{Character. Variable names in FinBIF native or Darwin Core
style.}
}
\value{
Character vector.
}
\description{
Convert FinBIF native variable names to Darwin Core style variable names or
vice versa.
}
\examples{

to_dwc("record_id", "date_time", "scientific_name")
}
