% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\title{Filtering FinBIF records}
\description{
Filters available for FinBIF records and occurrence data.
}
\section{Taxa}{

Filters related to taxa include:
\itemize{
\item \code{taxon_id} Character vector.
FinBIF taxon IDs. The functions \code{\link[=finbif_check_taxa]{finbif_check_taxa()}} and \code{\link[=finbif_taxa]{finbif_taxa()}}
can be used to search for taxon IDs.
\item \code{taxon_name} Character vector.
Filter based on taxon names (scientific or common) rather than IDs. If the
specified taxa are not found in the FinBIF taxonomy then matches are
attempted with the occurrence record names as originally supplied verbatim.
\item \code{quality_controlled_det} Logical.
If \code{TRUE} (default) use quality controlled taxonomic determinations. Or, if
\code{FALSE} use the originally recorded taxonomic determinations.
\item \code{subtaxa} Logical.
If \code{TRUE} (default) return records of all taxa belonging to the specified
taxon. Or, if \code{FALSE} only return records for exact matches to the
specified taxon (e.g., if a genus is specified do not return records of the
species belonging to genus only return records with individuals identified
as that genus only).
\item \code{invalid_taxa} Logical.
If \code{TRUE} (default) return records that have taxa found in the FinBIF
taxonomic database. Or, if \code{FALSE} limit records to only those with taxa
found in the FinBIF database.
\item \code{informal_group} Character vector.
Filter by informal taxonomic groups. Use the function
\code{\link[=finbif_informal_groups]{finbif_informal_groups()}} to see the informal taxonomic groups available
in FinBIF.
\item \code{informal_group_reported} Character.
\item \code{administrative_status} Character.
\item \code{redlist_status} Character.
\item \code{primary_habitat} Character.
\item \code{primary_secondary_habitat} Character.
\item \code{finnish}. Logical.
\item \code{invasive}. Logical
\item \code{taxon_rank}. Character.
}
}

\section{Location}{

Filters related to location of record include:
\itemize{
\item \code{country} Character.
\item \code{municipality} Character.
\item \code{province} Character.
\item \code{locality} Character.
\item \code{bird_assoc_area} Character.
\item \code{place_name_id} Character.
\item \code{coordinates_area} Coordinates.
\item \code{coordinates_center} Coordinates.
\item \code{coordinates_cell_*k} Coordinates.
\item \code{coordinates_cell_*k_center} Coordinates.
\item \code{coordinate_source} Character.
}
}

\section{Time}{

Filters related to time of record include:
\itemize{
\item \code{date_range_ymd} Date.
\item \code{date_range_ym} Date.
\item \code{date_range_d} Date.
\item \code{date_range_md} Date.
\item \code{last_import_date_min} Date.
\item \code{last_import_date_max} Date.
\item \code{first_import_date_min} Date.
\item \code{first_import_date_max} Date.
}
}

\section{Quality}{

Filters related to quality of record:
\itemize{
\item \code{collection_reliability} Integer.
\item \code{coordinate_accuracy_max} Integer.
\item \code{quality_issues}. Character.
\item \code{reliable} Logical.
\item \code{taxon_reliability} Character.
}
}

\section{Misc}{

Other filters:
\itemize{
\item \code{occurrence_type}, Character.
\item \code{not_occurrence_type} Character.
\item \code{form_id} Character.
\item \code{keywords} Character.
\item \code{collection} Character.
\item \code{not_collection} Character.
\item \code{subcollections} Logical.
\item \code{source} Character.
\item \code{record_basis} Character.
\item \code{super_record_basis} Character.
\item \code{life_stage} Character.
\item \code{sex} Character.
\item \code{invasive_control} Character.
\item \code{invasive_controlled} Logical.
\item \code{event_id} Character.
\item \code{document_id} Character.
\item \code{record_id} Character.
\item \code{individual_id} Character.
\item \code{abundance_min} Integer.
\item \code{abundance_max} Integer.
\item \code{type_specimen} Logical.
\item \code{is_native} Logical.
\item \code{wild_status} Character.
\item \code{is_breeding_location} Logical.
\item \code{has_document_media} Logical.
\item \code{has_event_media} Logical.
\item \code{has_record_media} Logical.
\item \code{has_media} Logical.
\item \code{editor_id} Character.
\item \code{observer_id} Character.
\item \code{editor_or_observer_id} Character.
\item \code{event_observer_name} Character.
\item \code{event_observer_id} Character.
\item \code{secure_reason} Character.
\item \code{secure_level} Character.
\item \code{secure} Logical.
\item \code{annotated} Logical.
\item \code{unidentified} Logical.
\item \code{taxon_census} Character.
\item \code{record_fact} List.
\item \code{event_fact} List.
\item \code{document_fact} List.
}
}

