% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif-package.R
\docType{package}
\name{finbif-package}
\alias{finbif}
\alias{finbif-package}
\title{finbif: Interface for the 'Finnish Biodiversity Information Facility' API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A programmatic interface to the 'Finnish Biodiversity Information
    Facility' ('FinBIF') API (<https://api.laji.fi>). 'FinBIF' aggregates
    Finnish biodiversity data from multiple sources in a single open access
    portal for researchers, citizen scientists, industry and government.
    'FinBIF' allows users of biodiversity information to find, access, combine
    and visualise data on Finnish plants, animals and microorganisms. The
    'finbif' package makes the publicly available data in 'FinBIF' easily
    accessible to programmers. Biodiversity information is available on taxonomy
    and taxon occurrence. Occurrence data can be filtered by taxon, time,
    location and other variables. The data accessed are conveniently
    preformatted for subsequent analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luomus/finbif}
  \item \url{https://luomus.github.io/finbif}
  \item Report bugs at \url{https://github.com/luomus/finbif/issues}
}

}
\author{
\strong{Maintainer}: William Morris \email{willi@mmorris.email}

Other contributors:
\itemize{
  \item LUOMUS - Finnish Museum of Natural History [copyright holder]
}

}
\keyword{internal}
