% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif-package.R
\docType{package}
\name{finbif-package}
\alias{finbif}
\alias{finbif-package}
\title{finbif: Finnish Biodiversity Information Facility R Interface}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

An R programming language interface to the FinBIF (Finnish
    Biodiversity Information Facility) API (api.laji.fi). FinBIF aggregates
    Finnish biodiversity data from multiple sources in a single open access
    portal for researchers, citizen scientists, industry and government. FinBIF
    allows users of biodiversity information to find, access, combine and
    visualise data on Finnish plants, animals and microorganisms. The finbif R
    package makes the publicly available data in FinBIF accessible from within
    R. Biodiversity information is available on taxonomy and taxon occurrence.
    Occurrence data can be filtered by taxon, time, location and other
    variables. The data accessed are conveniently preformatted for subsequent
    analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luomus/finbif}
  \item \url{https://luomus.github.io/finbif}
  \item Report bugs at \url{https://github.com/luomus/finbif/issues}
}

}
\author{
\strong{Maintainer}: William Morris \email{willi@mmorris.email}

}
\keyword{internal}
