% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif-package.R
\docType{package}
\name{finbif-package}
\alias{finbif}
\alias{finbif-package}
\title{finbif: Finnish Biodiversity Information Facility R Interface}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

An R interface to the FinBIF (Finnish Biodiversity Information 
    Facility) API (api.laji.fi).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bitbucket.org/luomus/finbif}
  \item \url{https://github.com/wkmor1/finbif}
  \item \url{https://finbif.netlify.com}
  \item Report bugs at \url{https://github.com/wkmor1/finbif/issues, https://bitbucket.org/luomus/finbif/issues}
}

}
\author{
\strong{Maintainer}: William Morris \email{willi@mmorris.email}

}
\keyword{internal}
