% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_clear_cache.R
\name{finbif_clear_cache}
\alias{finbif_clear_cache}
\title{Clear cache}
\usage{
finbif_clear_cache()
}
\description{
Remove cached API requests from FinBIF.
}
\examples{
\dontrun{

finbif_clear_cache()
}
}
