% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_collections.R
\name{finbif_collections}
\alias{finbif_collections}
\title{FinBIF collections}
\usage{
finbif_collections(subcollections = FALSE, supercollections = TRUE,
  nmin = 0)
}
\arguments{
\item{subcollections}{Logical. Return subcollection metadata of higher level
collections.}

\item{supercollections}{Logical. Return lowest level collection metadata.}

\item{nmin}{Integer. Filter collections by number of records. Only return
information on collections with greater than value specified. If \code{NA} then
return information on all collections.}
}
\value{
A data.frame.
}
\description{
Get information on collections in the FinBIF database.
}
\examples{
\dontrun{

# Get collection metadata
collecitons <- finbif_collections()
}
}
