% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_areas.R
\name{finbif_countries}
\alias{finbif_countries}
\alias{finbif_provinces}
\alias{finbif_municipalities}
\title{Countries, Provinces, and Municipalities}
\usage{
finbif_countries()

finbif_provinces()

finbif_municipalities()
}
\value{
A data.frame.
}
\description{
Display country, province or municipality names and codes used in the FinBIF
database.
}
\examples{
\dontrun{

# Display countries, provinces or municipalities
finbif_countries()
finbif_provinces()
finbif_municipalities()
}
}
