% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_informal_groups.R
\name{finbif_informal_groups}
\alias{finbif_informal_groups}
\alias{fb_informal_groups}
\title{FinBIF informal groups}
\usage{
finbif_informal_groups(
  group,
  limit = 5,
  quiet = FALSE,
  locale = getOption("finbif_locale"),
  cache = getOption("finbif_use_cache_metadata")
)
}
\arguments{
\item{group}{Character. Optional, if supplied only display this top-level
group and its subgroups.}

\item{limit}{Integer. The maximum number top-level informal groups (and their
sub-groups) to display.}

\item{quiet}{Logical. Return informal group names without displaying them.}

\item{locale}{Character. One of the supported two-letter ISO 639-1 language
codes. Current supported languages are English, Finnish and Swedish. For
data where more than one language is available the language denoted by
\code{locale} will be preferred while falling back to the other languages in the
order indicated above.}

\item{cache}{Logical or Integer. If \code{TRUE} or a number greater than zero,
then data-caching will be used. If not logical then cache will be
invalidated after the number of hours indicated by the argument.}
}
\value{
A character vector (invisibly).
}
\description{
Display the informal taxonomic groups used in the FinBIF database.
}
\examples{
\dontrun{

# Display the informal taxonomic groups used by FinBIF
finbif_informal_groups()

}
}
