% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_metadata.R
\name{finbif_metadata}
\alias{finbif_metadata}
\alias{fb_metadata}
\title{FinBIF metadata}
\usage{
finbif_metadata(which)
}
\arguments{
\item{which}{Character. Which category of metadata to display. If
unspecified, function returns the categories of metadata available.}
}
\value{
A data.frame.
}
\description{
Display metadata from the FinBIF database.
}
\examples{
\dontrun{

finbif_metadata("red_list")

}
}
