% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_occurrence.R
\name{finbif_occurrence}
\alias{finbif_occurrence}
\title{Download FinBIF occurrence records}
\usage{
finbif_occurrence(..., filters = NULL, fields, n = 10, page = 1)
}
\arguments{
\item{...}{Character vectors or list of character vectors. Taxa to download.}

\item{filters}{List of named character vectors. Filters to apply to records.}

\item{fields}{Character vector. Columns to return.}

\item{n}{Integer. How many records to download.}

\item{page}{Integer. Which page of records to start downloading from.}
}
\value{
A \code{data.frame}.
}
\description{
Download filtered occurence data from FinBIF as a \code{data.frame}.
}
