% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_occurrence.R
\name{finbif_occurrence}
\alias{finbif_occurrence}
\title{Download FinBIF occurrence records}
\usage{
finbif_occurrence(..., filters, fields, n = 10, page = 1,
  check_taxa = TRUE)
}
\arguments{
\item{...}{Character vectors or list of character vectors. Taxa of records
to download.}

\item{filters}{List of named character vectors. Filters to apply to records.}

\item{fields}{Character vector. Columns to return.}

\item{n}{Integer. How many records to download.}

\item{page}{Integer. Which page of records to start downloading from.}

\item{check_taxa}{Logical. Check first that taxa are in the FinBIF database.
If true only records that match known taxa (have a valid taxon ID) are
returned.}
}
\value{
A \code{data.frame}.
}
\description{
Download filtered occurrence data from FinBIF as a \code{data.frame}.
}
\examples{
\dontrun{

# Get recent occurrence data for taxon
finbif_occurrence("Cygnus cygnus")

# Specify the number of records
finbif_occurrence("Cygnus cygnus", n = 100)

# Get multiple taxa
finbif_occurrence("Cygnus cygnus", "Ursus arctos")

# Filter the record
finbif_occurrence(
  species = "Cygnus cygnus",
  filters = list(coordinateAccuracyMax = 100))
)

}
}
