% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_records.R
\name{finbif_records}
\alias{finbif_records}
\title{Get FinBIF records}
\usage{
finbif_records(filters, fields, n = 10, page = 1, count_only = FALSE,
  quiet = FALSE, cache = TRUE)
}
\arguments{
\item{filters}{List of named character vectors. Filters to apply to records.}

\item{fields}{Character vector. Columns to return. If not specified a default
set of commonly used fields will be used. Use \code{"default_fields"} as a
shortcut for this set.}

\item{n}{Integer. How many records to download.}

\item{page}{Integer. Which page of records to start downloading from.}

\item{count_only}{Logical. Only return the number of records available.}

\item{quiet}{Logical. Suppress the progress indicator for multipage
downloads.}

\item{cache}{Logical. Use cached data.}
}
\value{
A \code{finbif_api} or \code{finbif_api_list} object.
}
\description{
Download records from FinBIF.
}
\examples{
\dontrun{

# Get the last 100 records from FinBIF
finbif_records(n = 100)
}
}
