% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_records.R
\name{finbif_records}
\alias{finbif_records}
\title{Get FinBIF records}
\usage{
finbif_records(filters = NULL, fields, n = 10, page = 1)
}
\arguments{
\item{filters}{List of named character vectors. Filters to apply to records.}

\item{fields}{Character vector. Columns to return.}

\item{n}{Integer. How many records to download.}

\item{page}{Integer. Which page of records to start downloading from.}
}
\value{
A \code{finbif_api} object.
}
\description{
Download filtered records from FinBIF.
}
