% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_records.R
\name{finbif_records}
\alias{finbif_records}
\title{Get FinBIF records}
\usage{
finbif_records(filters, fields, n = 10, page = 1, count_only = FALSE)
}
\arguments{
\item{filters}{List of named character vectors. Filters to apply to records.}

\item{fields}{Character vector. Columns to return.}

\item{n}{Integer. How many records to download.}

\item{page}{Integer. Which page of records to start downloading from.}

\item{count_only}{Logical. Only return the number of records available.}
}
\value{
A \code{finbif_api} or \code{finbif_api_list} object.
}
\description{
Download records from FinBIF.
}
\examples{
\dontrun{

# Get the last 100 records from FinBIF
finbif_records(n = 100)
}
}
