% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_request_token.R
\name{finbif_request_token}
\alias{finbif_request_token}
\alias{fb_request_token}
\title{Get a FinBIF personal access token}
\usage{
finbif_request_token(email, quiet = FALSE)
}
\arguments{
\item{email}{Character. The email address to which to send the API access
token.}

\item{quiet}{Logical. Suppress messages.}
}
\value{
If an access token has already been set then \code{NULL} (invisibly) if
not then, invisibly, a \code{finbif_api} object containing the response from
the FinBIF server.
}
\description{
Have a personal access token for use with the FinBIF API sent to a specified
email address.
}
\examples{
\dontrun{

# Request a token for example@email.com
finbif_request_token("example@email.com")

}
}
