% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_taxa.R
\name{finbif_taxa}
\alias{finbif_taxa}
\title{Search the FinBIF taxa}
\usage{
finbif_taxa(name, n = 1, type = c("exact", "partial", "likely"))
}
\arguments{
\item{name}{Character. The name of a taxon to search for.}

\item{n}{Integer. Maximum number of matches to return.}

\item{type}{Character. Type of match to make. Must be one of \code{exact},
\code{partial} or \code{likely}.}
}
\value{
A \code{finbif_api} object.
}
\description{
Search the FinBIF database for taxon.
}
