% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finbif_update_cache.R
\name{finbif_update_cache}
\alias{finbif_update_cache}
\alias{fb_update_cache}
\title{Update cache}
\usage{
finbif_update_cache()
}
\description{
Update all cached API requests from FinBIF.
}
\examples{
\dontrun{

finbif_update_cache()

}
}
