% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{FinBIF record variables}
\description{
FinBIF record variables that can be selected in a finbif occurrence search.
}
\section{Identifiers}{

All identifiers are returned in the form of a URI. Identifiers include:
\itemize{
\item \code{occurrenceID} Character.
The ID of a record of organism's occurrence at a time and place.
\item \code{organismID} Character.
ID of an individual organism (e.g., a ringed bird that has been captured
multiple times will have a single \code{organismID} and multiple \code{occurrenceID}s
corresponding to each capture).
\item \code{eventID} Character.
Event ID. An event can contain one or more records (e.g., a survey of
plants at a particular location and time.)
\item \code{catalogNumber} Character.
Document ID. A set of events that share common metadata.
\item \code{formID} Character.
Form ID. The form used to create the document, event, record data.
\item \code{datasetID} Character.
Dataset ID. All documents, events, and records belong to a
collection/dataset (e.g., a museum collection, or the datasets collected
by a specific institution). Collections themselves can be part of a larger
(super)collection (e.g., all the collections at a specific museum). Only
the lowest level dataset ID for a record is returned. Use
\code{finbif_collections()} to explore the hierarchy of collections/datasets.
\item \code{sourceID} Character.
Source ID. The source of the collection's data.
}
}

\section{Taxa}{

Variables related to taxonomy of records include:
\itemize{
\item \code{taxonConceptID} Character.
The taxon ID in the form of a URI.
\item \code{originalTaxonConceptID} Character.
The taxon ID before (if any) annotation.
\item \code{annotatedTaxonConceptID} Character.
The new taxon ID if the record has had it's taxonomy annotated.
\item \code{verbatimTaxonConceptID} Character.
The taxon ID as originally reported by the record creator.
\item \code{scientificName} Character.
Scientific name of taxon.
\item \code{originalScientificName} Character.
The scientific name before (if any) annotation.
\item \code{scientificNameDisplay} Character.
Scientific name of taxon formatted for display (e.g., taxa with genus only
will be formatted as \emph{Genus sp.}).
\item \code{originalScientificNameDisplay} Character.
Scientific display name before (if any) annotation.
\item \code{vernacularName} Character.
Common (vernacular) name of taxon.
\item \code{originalVernacularName} Character.
Common name before (if any) annotation.
\item \code{verbatimIdentification} Character.
The name of the taxon as originally reported by the record creator.
\item \code{scientificNameItalicised} Logical.
Is the scientific name normally italicised (i.e., is the taxonomic rank
genus or below.)
\item \code{originalScientificNameItalicised} Logical.
Is the original scientific name normally italicised.
\item \code{scientificNameAuthorship} Character.
The authority for the taxon scientific name.
\item \code{originalScientificNameAuthorship} Character.
The authority for the taxon scientific name before (if any) annotation.
\item \code{verbatimScientificNameAuthorship} Character.
The authority of the taxon as originally reported by the record creator.
\item \code{taxonRank} Character.
The taxonomic rank of the taxon.
\item \code{originalTaxonRank} Character.
The taxonomic rank of the taxon before (if any) annotation.
\item \code{informalTaxonGroups} List.
The informal taxonomic groups that the taxon belongs to (e.g., birds) in
the form of URIs.
\item \code{originalInformalTaxonGroups} List.
The informal taxonomic groups that the taxon belonged to before (if any)
annotation.
\item \code{verbatimInformalTaxonGroups} List.
The informal taxonomic groups that the taxon belongs to as reported by the
record creator.
\item \code{taxonChecklist} Character.
The checklist (as a URI) that that taxon is found in.
\item \code{originalTaxonChecklist} Character.
The checklist (as a URI) that that taxon was found in before (if any)
annotation.
\item \code{taxonFinnishStatus} Logical.
Is the taxon considered Finnish. The definition of a Finnish taxon differs
by taxonomic group?
\item \code{originalFinnishStatus} Logical.
Was the taxon considered Finnish before (if any) annotation?
}
}

\section{Abundance, sex & life history}{

Variables related to abundance, sex and life history include:
\itemize{
\item \code{individualCount} Integer.
Number of individuals recorded.
\item \code{individualCountInterpreted} Integer.
Number of individuals recorded or inferred from the record. Note that many
records with \code{individualCount == 1L} only indicate the record is of
one individual and may not necessarily imply that this was the abundance at
that specified place and time (e.g., a preserved museum specimen consisting
of a single individual).
\item \verb{individualCount\{Female|Male\}} Integer.
Number of female or male individuals recorded.
\item \code{pairCount} Integer.
Number of mating pairs recorded.
\item \code{organismQuantity} Character.
The abundance as reported by the record creator.
\item \code{lifeStage} Character.
Life stage of individual(s) recorded.
\item \code{sex} Character.
Sex of individual(s) recorded.
}
}

\section{Location}{

Variables related to the location of records include:
\itemize{
\item \verb{decimal\{Latitude|Longitude\}} Numeric.
Coordinates (in
\href{https://spatialreference.org/ref/epsg/4326/}{WGS84} coordinate system)
of the central point of a bounding box encompassing the record's geographic
coverage.
\item \verb{decimal\{Latitude|Longitude\}EUREF} Numeric.
Coordinates (in
\href{https://spatialreference.org/ref/epsg/3067/}{EUREF} coordinate system)
of the central point of a bounding box encompassing the record's geographic
coverage.
\item \verb{decimal\{Latitude|Longitude\}\{Min|Max\}\{EUREF|YKJ|WGS84\}}. Numeric.
Vertices of a bounding box encompassing the record's geographic coverage.
Coordinates are available in
\href{https://spatialreference.org/ref/epsg/3067/}{EUREF},
\href{https://spatialreference.org/ref/epsg/2393/}{YKJ}, or
\href{https://spatialreference.org/ref/epsg/4326/}{WGS84}.
\item \code{coordinateUncertaintyInMeters} Integer.
The horizontal distance (in meters) from the record's given coordinates
describing the smallest circle containing the whole of the record's
location.
\item \code{georeferenceSources} Character.
Source of coordinates.
\item \code{footprintWKT} Character.
Well-Known Text (WKT) representation of the geographic shape defining the
location of the record in WGS84 coordinate systems.
\item \verb{footprintWKT\{_EUREF|_YKJ\}} Character.
Well-Known Text (WKT) representation of the geographic shape defining the
location of the record in either EUREF or YKJ coordinate systems.
\item \code{country} Character.
The country of the record's location.
\item \code{stateProvince} Character.
The administrative area directly below the level of country.
\item \code{bioStateProvince} Character.
For data from Finland FinBIF uses the concept of
\href{https://laji.fi/en/theme/emk}{Biogeographical Province}. See link for
details.
\item \code{county}. Character.
Administrative level below region
\item \code{higherGeography} Character.
Geographic place name that is at higher level than country.
\item \code{lineLengthMeters} Integer.
The length of linear locations (e.g., line transect surveys).
\item \code{areaMetersSquared} Integer.
The size of record's location in meters squared.
\item \code{isBreedingSite} Logical.
Whether or not the occurrence is recorded at a known breeding location.
\item \code{locationID} Character.
A location ID in the form of a URI.
\item \code{eventSection} Integer.
A numeric identifier for a sub-location of a location (e.g., a specific
part of a transect that undergoes repeated surveys.)
}
}

\section{Time}{

Variables related to time of record include:
\itemize{
\item \code{eventDateTime} POSIXct.
The date and time of the recording event. This variable is computed after
records are downloaded from FinBIF. Its timezone and accuracy can be
controlled see \code{finbif_occurrence()} for details.
\item \code{samplingEffort} Duration.
The duration of the recording event. This variable is computed after
records are downloaded from FinBIF.
\item \code{eventDateStart} Character.
The date the recording event began.
\item \code{eventDateEnd} Character.
The date the recording event ended.
\item \code{hourStart} Integer.
The hour (24 hour time) of the day the recording event began.
\item \code{hourEnd} Integer.
The hour (24 hour time) of the day the recording event ended.
\item \code{minuteStart} Integer.
The minute of the hour the recording event began.
\item \code{minuteEnd} Integer.
The minute of the hour the recording event started.
\item \code{startDayOfYear} Integer.
The ordinal day of the year the recording event began.
\item \code{endDayOfYear} Integer.
The ordinal day of the year the recording event ended
\item \code{seasonStart} Integer.
The day of the year the recording event began. A four digit number
indicating the day of the year in MMDD (\%m\%d) format.
\item \code{seasonEnd} Integer.
The day of the year the recording event ended. A four digit number
indicating the day of the year in MMDD (\%m\%d) format.
\item \code{century} Integer.
The century during which the recording event occurred (\code{NA} if the event
spans multiple centuries).
\item \code{decade} Integer.
The decade during which the  recording event occurred (\code{NA} if the event
spans multiple decades).
\item \code{year} Integer.
The year during which the recording event occurred (\code{NA} if the event spans
multiple years).
\item \code{month} Integer.
The month of the year during which the recording event occurred (\code{NA} if
the event spans multiple months).
\item \code{day} Integer.
The day of the month during which the recording event occurred (\code{NA} if the
event spans multiple days).
\item \code{eventDateTimeDisplay} Character.
Date and time of the recording event formatted for display.
\item \code{documentCreatedDate} Character.
The date the original data was created.
\item \code{available} Character.
The date the record was first loaded into the FinBIF database.
\item \code{modified} Character.
The most recent date the original data was modified.
\item \code{documentLoadedDate} Character.
The most recent date the record was loaded into the FinBIF database.
}
}

\section{Data restrictions}{

Variables related to restricted records include:
\itemize{
\item \code{informationRestricted} Logical.
Has the record been restricted in some way (e.g., geospatially aggregated).
\item \code{dataGeneralizations} Character.
What level of restriction has been applied to the record.
\item \code{informationWithheld} List.
List of reasons restriction has been applied.
}
}

\section{Data quality}{

Variables related to the quality of records include:
\itemize{
\item \code{hasIssues} Logical.
Are there any data quality issues associated with the record, its event or
document.
\item \code{identificationQualifer}
Reliability of the record's taxonomic identification as reported by the
original data author.
\item \verb{\{document|time|location|event|occurrence\}Issue} Character.
Issues with record associated with its document, time, location, event, or
the record itself.
\item \verb{\{document|time|location|event|occurrence\}IssueRemarks} Character.
Details about the issue.
\item \verb{\{document|time|location|event|occurrence\}IssueSource} Character.
Source determining the issue.
\item \code{requiresVerification} Logical.
Has the record been flagged for expert verification?
\item \code{requiresIdentification} Logical.
Has the record been flagged for expert identification?
\item \code{occurrenceReliability} Character.
Indication of the record's reliability.
\item \code{identificationVerificationStatus} Character.
Indication of the record's quality.
}
}

\section{Misc}{

Other variables:
\itemize{
\item \code{datasetName} Character.
Dataset name. All documents, events, and records belong to a
collection/dataset (e.g., a museum collection, or the datasets collected
by a specific institution). Collections themselves can be part of a larger
(super)collection (e.g., all the collections at a specific museum). Only
the lowest level collection name for a record is returned. Use
\code{finbif_collections()} to explore the hierarchy of collections/datasets.
\item \code{recordedByIDs} List.
List of observer identifiers for the record.
\item \code{identifiedBy} Character.
Person who determined the taxonomic identification of the record.
\item \code{basisOfRecord} Character.
The type of or method used to obtain the record.
\item \code{basisOfSuperRecord} Character.
Higher level type of or method used to obtain the record.
\item \code{typeStatus} Logical.
Whether or not the record is of a type specimen.
\item \code{occurrenceIsWild} Logical.
Whether or not the record is of a "wild" organism.
\item \code{license} Character.
The license of the data associated with the record.
\item \verb{\{document|event|occurrence\}Remarks} Character.
Notes associated with the document, event or record itself.
\item \verb{\{document|occurrence\}Keywords} List.
List of keywords associated with the document or record.
\item \code{occurrenceAnnotationCount} Integer.
How many annotations are associated with the record.
\item \code{materialSampleCount} Integer.
How many material samples (DNA extractions, etc., ...) are associated with
the record.
\item \verb{\{document|event|occurrence\}AssociatedMediaCount} Integer.
How many media items (images, audio, video, etc., ...) are associated with
the record's document, event or the record itself.
}
}

