% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{FinBIF record variables}
\description{
FinBIF record variables that can be selected in a finbif occurrence search.
}
\section{Identifiers}{

All identifiers are returned in the form of a URI. Identifiers include:
\itemize{
\item \code{record_id}
Record ID. The ID of a record of organism's occurrence at a time and
place.
\item \code{individual_id}
ID of an individual organism (e.g., a ringed bird that has been captured
multiple times will have a single \code{individual_id} and multiple \code{record_id}s
corresponding to each capture).
\item \code{event_id}
Event ID. An event can contain one or more records (e.g., a survey of
plants at a particular location and time.)
\item \code{document_id}
Document ID. A set of events that share common metadata.
\item \code{form_id}
Form ID. The form used to create the document, event, record data.
\item \code{collection_id}
Collection ID. All documents, events, and records belong to a collection (
e.g., a museum collection, or the datasets collected by a specific
institution). Collections themselves can be part of a larger
(super)collection (e.g., all the collections at specific museum). Only the
lowest level collection ID for a record is returned. Use
\code{finbif_collections()} to explore the hierarchy of collections.
\item \code{source_id}
Source ID. The source of the collection's data.
}
}

\section{Taxa}{

Variables related to taxonomy of records include:
\itemize{
\item \code{taxon_id} Character.
The taxon ID in the form of a URI.
\item \code{orig_taxon_id} Character.
The taxon ID (above) before (if any) annotation.
\item \code{annotated_taxon_id} Charactio
The new taxon ID if the record has had it's taxonomy annotated.
\item \code{reported_taxon_id}
The taxon ID reported by the original data provider (may not exist).
\item \code{scientific_name} Character.
Scientific name of taxon.
\item \code{orig_scientific_name}
\item \code{scientific_display_name}
\item \code{orig_scientific_display_name}
\item \code{vernacular_name}
\item `orig_name_vernacular.
\item \code{taxon_name_cursive}
\item \code{orig_taxon_cursive}
\item \code{taxon_verbatim}
\item \code{scientific_name_author}
\item \code{orig_scientific_name_author}
\item \code{author_verabtum}
\item \code{taxon_rank} Character.
Taxonomic rank of the taxon (in the form of a URI).
\item \code{orig_taxon_rank}
\item \code{informal_groups}
\item \code{orig_informal_groups}
\item \code{reported_informal_group}
\item \code{taxon_checklist}
\item \code{orig_taxon_checklist}
\item \code{taxon_finnish}
\item \code{orig_taxon_finnish}
\item \code{is_unidentifiable}
}
}

\section{Abundance, sex & life history}{

Variables related to abundance, sex and life history include:
\itemize{
\item \code{abundance} Integer.
Number of individuals recorded or inferred from the record. Note that many
records with \code{abundance = 1} only indicate the record is of one individual
and may not necessarily imply that this was the abundance at that specified
place and time (e.g., a preserved museum specimen consisting of a single
individual).
\item \code{{female|male}_abundance}
\item \code{pair_abundance}
\item \code{abundance_verbatum}
\item \code{life_stage}
\item \code{sex}
}
}

\section{Location}{

Variables related to the location of record include:
\itemize{
\item \code{{lat|lon}_wgs84}
\item \code{{lat|lon}_{min|max}_{euref|kkj|wgs84}}
\item \code{{lat|lon}_{005|01|05|1}_wgs84}
\item \code{{lat|lon}_{1|10|50|100}_kkj}
\item \code{{lat|lon}_{1|10|50|100}_center_kkj}
\item \code{wkt_{euref|kkj|wgs84}}
\item \code{coordinates_verbatim}
\item \code{coordinate_accuracy}
\item \code{coordinates_source}
\item \code{country}
\item \code{country_id}
\item \code{country_source}
\item \code{country_verbatim}
\item \code{province}
\item \code{province_id}
\item \code{province_ids}
\item \code{province_source}
\item \code{province_verbatum}
\item \code{province_verbatim2}
\item \code{municipality}
\item \code{municipality_id}
\item \code{municipality_ids}
\item \code{municipality_source}
\item \code{municipality_verbatum}
\item \code{locality}
\item \code{location_id}
\item \code{higher_geography}
\item \code{line_length_m}
\item \code{area_m2}
\item \code{is_breeding_location}
}
}

\section{Time}{

Variables related to time of record include:
\itemize{
\item \code{date_start}
\item \code{date_end}
\item \code{hour_start}
\item \code{hour_end}
\item \code{minute_start}
\item \code{minute_end}
\item \code{ordinal_day_start}
\item \code{ordinal_day_end}
\item \code{season_start}
\item \code{season_end}
\item \code{formatted_date_time}
\item \code{century}
\item \code{decade}
\item \code{year}
\item \code{month}
\item \code{day}
\item \code{date_created}
\item \code{first_load_date}
\item \code{modified_date}
\item \code{load_date}
}
}

\section{Data restrictions}{

Variables related to restricted records include:
\itemize{
\item \code{restricted}
\item \code{restriction_level}
\item \code{restriction_reasons}
\item \code{conservation_reason_restricted}
\item \code{custom_reason_restricted}
\item \code{data_quarantine_period_reason_restricted}
}
}

\section{Data quality}{

Variables related to the quality of records include:
\itemize{
\item \code{any_issues}
\item \code{record_reliable}
\item \code{document_reliablity}
\item \code{taxon_reliability}
\item \code{taxon_reliability_message}
\item \code{taxon_reliability_source}
\item \code{reported_taxon_confidence}
\item \code{{document|time|location|event|record}_issue}
\item \code{{document|time|location|event|record}_issue_message}
\item \code{{document|time|location|event|record}_issue_source}
}
}

\section{Media}{

Variables related to media (images, audio, etc.,) associated with records
include:
\itemize{
\item \code{{document|event|record}_media_count}
\item \code{{document|event|record}_media_author}
\item \code{{document|event|record}_media_caption}
\item \code{{document|event|record}_media_copyright}
\item \code{{document|event|record}_media_url}
\item \code{{document|event|record}_media_license_abbr}
\item \code{{document|event|record}_media_license_id}
\item \code{{document|event|record}_media_type}
\item \code{{document|event|record}_thumbnail_square}
\item \code{{document|event|record}_thumbnail}
}
}

\section{Misc}{

Other variables:
\itemize{
\item \code{observers_name}
\item \code{observers_id}
\item \code{observers_user_id}
\item \code{observer_user_ids}
\item \code{editor_name}
\item \code{editor_id}
\item \code{editor_user_id}
\item \code{determiner}
\item \code{team}
\item \code{{document|event|record}_fact_decimal}
\item \code{{document|event|record}_fact_content}
\item \code{{document|event|record}_fact_integer}
\item \code{{document|event|record}_fact_value}
\item \code{invasive_control_effectiveness}
\item \code{invasive_control}
\item \code{record_notes}
\item \code{record_basis}
\item \code{reference_publication}
\item \code{superrecord_basis}
\item \code{type_specimen}
\item \code{record_order}
\item \code{wild_status}
\item \code{editor_user_ids}
\item \code{keywords}
\item \code{license}
\item \code{document_notes}
\item \code{partial}
\item \code{event_order}
\item \code{event_notes}
\item \code{taxon_cenus_id}
\item \code{taxon_census_type}
\item \code{record_annotation_count}
\item \code{{document|record}_annotation_person_id}
\item \code{{document|record}_annotation_person_name}
\item \code{{document|record}_annotation_system_id}
\item \code{{document|record}_annotation_system_name}
\item \code{{document|record}_annotation_class}
\item \code{{document|record}_annotation_created}
\item \code{{document|record}_annotation_id}
\item \code{{document|record}_annotation_invasive_control_effectiveness}
\item \code{{document|record}_annotation_notes}
\item \code{{document|record}_annotation_opinion}
\item \code{{document|record}_annotation_root_id}
\item \code{{document|record}_annotation_target_id}
\item \code{{document|record}_annotation_type}
\item \code{keywords}
\item \code{sample_count}
\item \code{sample_collection_id}
\item \code{sample_fact_decimal}
\item \code{sample_fact_content}
\item \code{sample_fact_integer}
\item \code{sample_fact_value}
\item \code{sample_keywords}
\item \code{sample_multiple}
\item \code{sample_notes}
\item \code{sample_quality}
\item \code{sample_id}
\item \code{sample_order}
\item \code{sample_status}
\item \code{samples_type}
}
}

