% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{FinBIF record variables}
\description{
FinBIF record variables that can selected.
}
\section{Misc}{

Other variables:
\itemize{
\item \code{scientific_name} Character.
Scientific name of taxon.
\item \code{taxon_rank} Character.
Taxonomic rank of the taxon (in the form of a URI).
\item \code{abundance} Integer.
Number of individuals recorded or inferred from the record. Note that many
records with \code{abundance = 1} only indicate the record is of one individual
and may not necessarily imply that this was the abundance at that specified
place and time (e.g., a preserved museum specimen consisting of a single
individual).
\item \code{abundance_string}
\item \code{country}
\item \code{province}
\item \code{municipality}
\item \code{lat_wgs84}
\item \code{lon_wgs84}
\item \code{wkt_wgs84}
\item \code{line_length_m}
\item \code{area_m2}
\item \code{date_start}
\item \code{date_end}
\item \code{hour_start}
\item \code{hour_end}
\item \code{minute_start}
\item \code{minute_end}
\item \code{record_id}
\item \code{individual_id}
\item \code{event_id}
\item \code{collection_id}
\item \code{annotation_count}
\item \code{annotation_person}
\item \code{annotation_person_name}
\item \code{annotation_system}
\item \code{annotation_system_name}
\item \code{annotation_class}
\item \code{annotation_created}
\item \code{annotation_id}
\item \code{annotation_invasive_control_effectiveness}
\item \code{annotation_notes}
\item \code{annotation_opinion}
\item \code{annotation_type}
\item \code{author}
\item \code{is_breeding_site}
\item \code{determiner}
\item \code{record_fact_decimal}
\item \code{record_fact}
\item \code{record_fact_integer}
\item \code{record_fact_value}
\item \code{female_abundance}
\item \code{male_abundance}
\item \code{annotated_taxon_id}
\item \code{pair_abundance}
\item \code{is_unidentifiable}
\item \code{invasive_control_effectiveness}
\item \code{invasive_control}
\item \code{life_stage}
\item \code{orig_taxon_checklist}
\item \code{orig_taxon_cursive}
\item \code{orig_taxon_finnish}
\item \code{orig_taxon_id}
\item \code{orig_informal_groups}
\item \code{orig_qname}
\item \code{orig_taxon_scientific_name}
\item \code{orig_scientific_name_author}
\item \code{orig_scientific_display_name}
\item \code{orig_taxon_rank}
\item \code{orig_name_vernacular}
\item \code{taxon_checklist}
\item \code{taxon_name_cursive}
\item \code{taxon_finnish}
\item \code{taxon_id}
\item \code{informal_groups}
\item \code{qname}
\item \code{scientific_name_author}
\item \code{scientific_display_name}
\item \code{vernacular_name}
\item \code{record_media_author}
\item \code{record_media_caption}
\item \code{record_media_copyright}
\item \code{record_media_url}
\item \code{record_media_license_abbr}
\item \code{record_media_license_id}
\item \code{record_media_type}
\item \code{record_thumbnail_square}
\item \code{record_thumbnail}
\item \code{record_media_count}
\item \code{record_notes}
\item \code{any_issues}
\item \code{record_issue}
\item \code{record_issue_message}
\item \code{record_issue_source}
\item \code{record_reliable}
\item \code{taxon_issue_message}
\item \code{taxon_reliability}
\item \code{taxon_issue_source}
\item \code{record_basis}
\item \code{reference_publication}
\item \code{reported_informal_group}
\item \code{reported_taxon_confidence}
\item \code{reported_taxon_id}
\item \code{sex}
\item \code{superrecord_basis}
\item \code{taxon_verbatim}
\item \code{type_specimen}
\item \code{record_order}
\item \code{wild_status}
\item \code{conservation_reason_secured}
\item \code{date_created}
\item \code{custom_reason_secured}
\item \code{data_quarantine_period_reason_secured}
\item \code{document_id}
\item \code{editor_user_ids}
\item \code{document_fact_decimal}
\item \code{document_fact_content}
\item \code{document_fact_integer}
\item \code{document_fact_value}
\item \code{first_load_date}
\item \code{form_id}
\item \code{keywords}
\item \code{license}
\item \code{editor_name}
\item \code{editor_id}
\item \code{editor_user_id}
\item \code{load_date}
\item \code{document_media_author}
\item \code{document_media_caption}
\item \code{document_media_copyright}
\item \code{document_media_url}
\item \code{document_media_license_abbr}
\item \code{document_media_license_id}
\item \code{document_media_type}
\item \code{document_thumbnail_square}
\item \code{document_thumbnail}
\item \code{document_media_count}
\item \code{modified_date}
\item \code{bird_assoc_area}
\item \code{bird_assoc_area_id}
\item \code{location_collection_id}
\item \code{location_id}
\item \code{location_municipality}
\item \code{location_municipality_id}
\item \code{location_name}
\item \code{location_lat_kkj}
\item \code{location_lon_kkj}
\item \code{locationid}
\item \code{document_notes}
\item \code{partial}
\item \code{document_issue}
\item \code{document_issue_message}
\item \code{document_issue_source}
\item \code{document_reliablity}
\item \code{secured_level}
\item \code{secured_reasons}
\item \code{secured}
\item \code{source_id}
\item \code{province_verbatim}
\item \code{century}
\item \code{day}
\item \code{ordinal_day_start}
\item \code{ordinal_day_end}
\item \code{decade}
\item \code{lat_max_euref}
\item \code{lat_min_euref}
\item \code{lon_max_euref}
\item \code{lon_min_euref}
\item \code{wkt_euref}
\item \code{month}
\item \code{season_start}
\item \code{season_end}
\item \code{lat_max_wgs84}
\item \code{lat_min_wgs84}
\item \code{lon_max_wgs84}
\item \code{lon_min_wgs84}
\item \code{lat_005_wgs84}
\item \code{lon_005_wgs84}
\item \code{lat_01_wgs84}
\item \code{lon_01_wgs84}
\item \code{lat_05_wgs84}
\item \code{lon_05_wgs84}
\item \code{lat_1_wgs84}
\item \code{lon_1_wgs84}
\item \code{year}
\item \code{lat_max_kkj}
\item \code{lat_min_kkj}
\item \code{lon_max_kkj}
\item \code{lon_min_kkj}
\item \code{lat_100_kkj}
\item \code{lon_100_kkj}
\item \code{lat_100_center_kkj}
\item \code{lon_100_center_kkj}
\item \code{lat_10_kkj}
\item \code{lon_10_kkj}
\item \code{lat_10_center_kkj}
\item \code{lon_10_center_kkj}
\item \code{lat_1_kkj}
\item \code{lon_1_kkj}
\item \code{lat_1_center_kkj}
\item \code{lon_1_center_kkj}
\item \code{lat_50_kkj}
\item \code{lon_50_kkj}
\item \code{lat_50_center_kkj}
\item \code{lon_50_center_kkj}
\item \code{wkt_kkj}
\item \code{coordinates_verbatim}
\item \code{country_verbatim}
\item \code{date_time}
\item \code{event_fact_decimal}
\item \code{event_fact}
\item \code{event_fact_integer}
\item \code{event_fact_value}
\item \code{event_order}
\item \code{higher_geography}
\item \code{province_id}
\item \code{province_ids}
\item \code{coordinate_accuracy}
\item \code{country_id}
\item \code{municipalities_ids}
\item \code{municipality_id}
\item \code{province_source}
\item \code{coordinates_source}
\item \code{country_source}
\item \code{municipality_source}
\item \code{observers_name}
\item \code{observers_id}
\item \code{observers_user_id}
\item \code{locality}
\item \code{event_media_author}
\item \code{event_media_caption}
\item \code{event_media_copyright}
\item \code{event_media_url}
\item \code{event_media_license_abbr}
\item \code{event_media_license_id}
\item \code{event_media_type}
\item \code{event_thumbnail_square}
\item \code{event_thumbnail}
\item \code{event_media_count}
\item \code{municipality_verbatum}
\item \code{province_verbatum2}
\item \code{event_notes}
\item \code{observer_user_ids}
\item \code{event_issue}
\item \code{event_issue_message}
\item \code{event_issue_source}
\item \code{location_issue}
\item \code{location_issue_message}
\item \code{location_issue_source}
\item \code{time_issue}
\item \code{time_issue_message}
\item \code{time_issue_source}
\item \code{taxon_cenus_id}
\item \code{taxon_census_type}
\item \code{team}
}
}

