% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{FinBIF record variables}
\description{
FinBIF record variables that can be selected in a finbif occurrence search.
}
\section{Identifiers}{

All identifiers are returned in the form of a URI. Identifiers include:
\itemize{
\item \code{record_id} Character.
The ID of a record of organism's occurrence at a time and
place.
\item \code{individual_id} Character.
ID of an individual organism (e.g., a ringed bird that has been captured
multiple times will have a single \code{individual_id} and multiple \code{record_id}s
corresponding to each capture).
\item \code{event_id} Character.
Event ID. An event can contain one or more records (e.g., a survey of
plants at a particular location and time.)
\item \code{document_id} Character.
Document ID. A set of events that share common metadata.
\item \code{form_id} Character.
Form ID. The form used to create the document, event, record data.
\item \code{collection_id} Character.
Collection ID. All documents, events, and records belong to a collection
(e.g., a museum collection, or the datasets collected by a specific
institution). Collections themselves can be part of a larger
(super)collection (e.g., all the collections at a specific museum). Only
the lowest level collection ID for a record is returned. Use
\code{finbif_collections()} to explore the hierarchy of collections.
\item \code{source_id} Character.
Source ID. The source of the collection's data.
}
}

\section{Taxa}{

Variables related to taxonomy of records include:
\itemize{
\item \code{taxon_id} Character.
The taxon ID in the form of a URI.
\item \code{orig_taxon_id} Character.
The taxon ID before (if any) annotation.
\item \code{annotated_taxon_id} Character.
The new taxon ID if the record has had it's taxonomy annotated.
\item \code{reported_taxon_id} Character.
The taxon ID as originally reported by the record creator.
\item \code{scientific_name} Character.
Scientific name of taxon.
\item \code{orig_scientific_name} Character.
The scientific name before (if any) annotation.
\item \code{scientific_display_name} Character.
Scientific name of taxon formatted for display (e.g., taxa with genus only
will be formatted as \emph{Genus sp.}).
\item \code{orig_scientific_display_name} Character.
Scientific display name before (if any) annotation.
\item \code{common_name} Character.
Common name of taxon.
\item \code{orig_common_name} Character.
Common name before (if any) annotation.
\item \code{reported_name} Character.
The name of the taxon as originally reported by the record creator.
\item \code{scientific_name_italicised} Logical.
Is the scientific name normally italicised (i.e., is the taxonomic rank
genus or below.)
\item \code{orig_scientific_name_italicised} Logical.
Is the original scientific name normally italicised.
\item \code{scientific_name_author} Character.
The authority for the taxon scientific name.
\item \code{orig_scientific_name_author} Character.
The authority for the taxon scientific name before (if any) annotation.
\item \code{reported_author} Character.
The authority of the taxon as originally reported by the record creator.
\item \code{taxon_rank} Character.
The taxonomic rank of the taxon (in the form of a URI).
\item \code{orig_taxon_rank} Character.
The taxonomic rank of the taxon (in the form of a URI) before (if any)
annotation.
\item \code{informal_groups} Character vector.
The informal taxonomic groups that the taxon belongs to (e.g., birds) in
the form of URIs.
\item \code{orig_informal_groups} Character vector.
The informal taxonomic groups that the taxon belonged to before (if any)
annotation.
\item \code{reported_informal_groups} Character vector.
The informal taxonomic groups that the taxon belongs to as reported by the
record creator.
\item \code{taxon_checklist} Character.
The checklist (as a URI) that that taxon is found in.
\item \code{orig_taxon_checklist} Character.
The checklist (as a URI) that that taxon was found in before (if any)
annotation.
\item \code{taxon_finnish} Logical.
Is the taxon considered Finnish. The definition of a Finnish taxon differs
by taxonomic group?
\item \code{orig_taxon_finnish} Logical.
Was the taxon considered Finnish before (if any) annotation?
\item \code{is_unidentifiable} Logical.
Is the record unable to be identified?
}
}

\section{Abundance, sex & life history}{

Variables related to abundance, sex and life history include:
\itemize{
\item \code{abundance} Integer.
Number of individuals recorded or inferred from the record. Note that many
records with \code{abundance = 1} only indicate the record is of one individual
and may not necessarily imply that this was the abundance at that specified
place and time (e.g., a preserved museum specimen consisting of a single
individual).
\item \code{{female|male}_abundance} Integer.
Number of female or male individuals recorded.
\item \code{pair_abundance} Integer.
Number of mating pairs recorded.
\item \code{abundance_verbatim} Character.
The abundance as reported by the record creator.
\item \code{life_stage} Character.
Life stage of individual(s) recorded.
\item \code{sex} Character.
Sex of individual(s) recorded.
}
}

\section{Location}{

Variables related to the location of records include:
\itemize{
\item \code{{lat|lon}_wgs84} Numeric.
Coordinates (in \href{https://spatialreference.org/ref/epsg/wgs-84/}{WGS84}
coordinate system) of the central point of a bounding box encompassing the
record's geographic coverage.
\item \code{{lat|lon}_{min|max}_{euref|kkj|wgs84}}. Numeric.
Vertices of a bounding box encompassing the record's geographic coverage.
Coordinates are available in
\href{https://spatialreference.org/ref/epsg/etrs89-etrs-tm35fin/}{EUREF},
\href{https://spatialreference.org/ref/epsg/2393/}{KKJ}, or
\href{https://spatialreference.org/ref/epsg/wgs-84/}{WGS84}.
\item \code{coordinates_uncertainty} Integer.
The horizontal distance (in meters) from the record's given coordinates
describing the smallest circle containing the whole of the record's
location.
\item \code{coordinates_source}
\item \code{footprint_{euref|kkj|wgs84}} Character.
Well-Known Text (WKT) representation of the geographic shape defining the
location of the record in either EUREF, KKJ or WGS84 coordinate systems.
\item \code{country} Character.
The country of the record's location.
\item \code{province} Character.
The administrative area directly below the level of country. For data from
Finland FinBIF uses the concept of
\href{https://laji.fi/en/theme/emk}{Biogeographical Province}. See link for
details.
\item \code{municipality}. Character.
Administrative level below province.
\item \code{higher_geography} Character.
Geographic place name that is at higher level than country.
\item \code{line_length_m} Integer.
The length of linear locations (e.g., line transect surveys).
\item \code{area_m2}
The size of record's location.
\item \code{is_breeding_location} Logical.
Whether or not the occurrence is recorded at a known breeding location.
}
}

\section{Time}{

Variables related to time of record include:
\itemize{
\item \code{date_start}
\item \code{date_end}
\item \code{hour_start}
\item \code{hour_end}
\item \code{minute_start}
\item \code{minute_end}
\item \code{ordinal_day_start}
\item \code{ordinal_day_end}
\item \code{season_start}
\item \code{season_end}
\item \code{formatted_date_time}
\item \code{century}
\item \code{decade}
\item \code{year}
\item \code{month}
\item \code{day}
\item \code{date_created}
\item \code{first_load_date}
\item \code{modified_date}
\item \code{load_date}
}
}

\section{Data restrictions}{

Variables related to restricted records include:
\itemize{
\item \code{restricted}
\item \code{restriction_level}
\item \code{restriction_reasons}
\item \code{conservation_reason_restricted}
\item \code{custom_reason_restricted}
\item \code{data_quarantine_period_reason_restricted}
}
}

\section{Data quality}{

Variables related to the quality of records include:
\itemize{
\item \code{any_issues}
\item \code{record_reliable}
\item \code{document_reliablity}
\item \code{taxon_reliability}
\item \code{taxon_reliability_message}
\item \code{taxon_reliability_source}
\item \code{reported_taxon_confidence}
\item \code{{document|time|location|event|record}_issue}
\item \code{{document|time|location|event|record}_issue_message}
\item \code{{document|time|location|event|record}_issue_source}
}
}

\section{Media}{

Variables related to media (images, audio, etc.,) associated with records
include:
\itemize{
\item \code{{document|event|record}_media_count}
\item \code{{document|event|record}_media_author}
\item \code{{document|event|record}_media_caption}
\item \code{{document|event|record}_media_copyright}
\item \code{{document|event|record}_media_url}
\item \code{{document|event|record}_media_license_abbr}
\item \code{{document|event|record}_media_license_id}
\item \code{{document|event|record}_media_type}
\item \code{{document|event|record}_thumbnail_square}
\item \code{{document|event|record}_thumbnail}
}
}

\section{Misc}{

Other variables:
\itemize{
\item \code{observers_ids}
\item \code{determiner}
\item \code{record_notes}
\item \code{record_basis}
\item \code{reference_publication}
\item \code{superrecord_basis}
\item \code{type_specimen}
\item \code{record_order}
\item \code{wild_status}
\item \code{license}
\item \code{document_notes}
\item \code{partial}
\item \code{event_order}
\item \code{event_notes}
\item \code{taxon_cenus_id}
\item \code{taxon_census_type}
\item \code{record_annotation_count}
\item \code{{document|record}_keywords}
\item \code{sample_count}
}
}

